+++ app/Http/Controllers/Admin/UserController.php
@@ -2,27 +2,30 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Models\User;
+
 use Illuminate\Http\Request;
 
-use App\Http\Controllers\Controller;
-
-use App\Models\User;
 use Spatie\Permission\Models\Role;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
+use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
-    private $totalPage = 10;
+    private $totalPage = 30;
 
     public function __construct()
     {
-        $this->middleware(['permission:users']);
+        // $this->middleware(['permission:user_access']);
     }
 
     public function index()
     {
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN);
+
         $users = User::all();
-        // $users = User::paginate();
         // $users = User::paginate($this->totalPage);
 
         return view('admin.users.index', compact('users'));
@@ -30,6 +33,8 @@ class UserController extends Controller
 
     public function create()
     {
+        abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN);
+
         $roles = Role::get();
         $permissions = Permission::all();
 
@@ -38,6 +43,8 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
+        abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN);
+
         $this->validate($request, [
             'name'=>'required|max:120',
             'email'=>"required|email|unique:users,email,id",
@@ -54,24 +61,24 @@ class UserController extends Controller
 
             return redirect()
                     ->route('admin.users.index')
-                    ->with(
-                        'success',
-                        'User ' . $user->name . ' added.'
-                    );
+                    ->with('success', "Usuário {$user->name} cadastrado com sucesso.");
         }
 
         return redirect()
             ->back()
-            ->with('error', 'User '. $user->name.' failured!');
+            ->with('error', "Erro ao cadastrar o Usuário {$user->name}.");
     }
 
     public function show($id)
     {
+        abort_if(Gate::denies('user_show'), Response::HTTP_FORBIDDEN);
         return redirect('users');
     }
 
     public function edit($id)
     {
+        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN);
+
         $user = User::findOrFail($id);
         $roles = Role::get();
         $permissions = Permission::all();
@@ -81,6 +88,8 @@ class UserController extends Controller
 
     public function update(Request $request, $id)
     {
+        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN);
+
         $this->validate($request, [
             'name'=>'required|max:120',
             'email'=>"required|email|unique:users,email,{$id},id",
@@ -104,29 +113,28 @@ class UserController extends Controller
 
             return redirect()
                     ->route('admin.users.index')
-                    ->with('success', 'User successfully edited.');
+                    ->with('success', "Usuário {$user->name} editado com sucesso.");
         }
 
         return redirect()
                 ->back()
-                ->with('error', 'Falha ao atualizar o User');
+                ->with('error', "Erro ao editar o Usuário {$user->name}.");
     }
 
     public function destroy($id)
     {
+        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN);
+
         $user = User::findOrFail($id);
         $delete = $user->delete();
         if ($delete) {
             return redirect()->route('admin.users.index')
-                ->with(
-                    'success',
-                    'User successfully deleted.'
-                );
+                ->with('success', "Usuário {$user->name} excluído com sucesso.");
         }
 
         return redirect()
             ->back()
-            ->with('error', 'Falha ao deletar o Usuário');
+            ->with('error', "Erro ao excluir o Usuário {$user->name}.");
     }
 
     /**
@@ -136,12 +144,17 @@ class UserController extends Controller
      */
     public function massDestroy(Request $request)
     {
+        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN);
+        
         if ($request->input('ids')) {
             $users = User::whereIn('id', $request->input('ids'))->get();
 
             foreach ($users as $user) {
                 $user->delete();
             }
+
+            return redirect()->route('admin.users.index')
+                ->with('success', 'Usuários excluídos com sucesso.');
         }
     }
 }
